ECDR-67: Moved mapper to its own bundle and made other fixes and tweaks.
