+++ apps/cdr-app/src/main/resources/features.xml
@@ -27,6 +27,7 @@
         <feature>cdr-ddf-registry-core</feature>
         <bundle>mvn:net.di2e.ecdr.transformer/cdr-transformer-mapper/${project.version}</bundle>
         <bundle>mvn:net.di2e.ecdr.libs/cdr-rest-search-commons/${project.version}</bundle>
+        <bundle>mvn:net.di2e.ecdr.endpoint/cdr-endpoint-sortmapper/${project.version}</bundle>
         <bundle>mvn:net.di2e.ecdr.libs/cdr-libs-result-relevance-normalizer/${project.version}</bundle>
         <bundle>mvn:net.di2e.ecdr.federation/cdr-federation-strategies/${project.version}</bundle>
         <bundle>mvn:net.di2e.ecdr.endpoint/cdr-rest-search-endpoint/${project.version}</bundle>
